# Don't allocate lines during testing in constructor.