--- a/Elements/test/PolygonTests.cs
+++ b/Elements/test/PolygonTests.cs
@@ -1861,8 +1861,7 @@ namespace Elements.Geometry.Tests
                                     (8.01, 26.90062, 0),
                                     (0, 26.90062, 0),
                                     (0, 0, 0),
-                                    (8.01, 0, 0),
-                                    (8.01, 6, 0));
+                                    (8.01, 0, 0));
             var line = new Line((2.930087, 15.546126, 0), (2.830087, 15.546126, 0));
             var extension = line.ExtendTo(pgon.Segments());
             Assert.True(extension.Direction() == line.Direction());